WARNING: BREAKING COMMIT
I was going to make this 1.5.3, but then I realized I shouldn't hide something this big behind a sub-minor release. For reasons of having the module included in the module list on CPAN, making everything be a bit neater, and for all-around happiness, I've moved Perlwikipedia into the preexisting MediaWiki:: namespace. The module is now called MediaWiki::Bot.

You can also now use "PWP" instead of "MediaWiki::Bot" in your code. The distribution on CPAN and in the svn repo includes an alias.

Plugins are now loaded automatically if installed. This means you don't have to do as much arcane magic. The plugins will be listed under MediaWiki::Bot::Plugin:: in CPAN and are not currently in svn. If you have a task that you would like to make into a plugin and publish on CPAN, please get in touch with me at DCOLLINS@cpan.org

Plugin notes since the last release:

Added the ImageTester plugin to CPAN. Purpose is to analyze and tag copyrighted images which do not follow en.wikipedia's policies

Added the CUP plugin to CPAN as a basic example of what a plugin does. This is used by the bot that will run the WikiCup stats, and these methods can be called just like any other methods in the main module



git-svn-id: https://perlwikipedia.googlecode.com/svn@179 6dbc9f61-ac24-0410-afbc-9547d35e9fc7