[BREAKING] Issue 31: Return undef instead of error codes
This continues the saga of getting rid of returned error codes.

Create a private sub to handle errors from the API which returns
undef. Error codes are available in $bot->{'error'}->{'code'}.
Error-checking in calling code can either be ignored as previously,
or can remain with just a change in the error condition.

Misc:
 * Add a test for error handling.
 * Update MANIFEST accordingly
 * Update t/11_get_pages_in_namespace.t accordingly

git-svn-id: https://perlwikipedia.googlecode.com/svn@251 6dbc9f61-ac24-0410-afbc-9547d35e9fc7