Fix input handling for getting contents of categories
This attempts to avoid a network round-trip to check the localized
namespace name for ns14 (Category) by allowing 'Category' to be
used (it is the canonical namespace name). Where needed, check if
the prefix provided is the localized category namespace name. If
not, then prepend it. Note that when passing localized namespace
names which use unicode, you'll need to pass `skip_encoding => 1`
in the options_hashref.

Resolves Issue 97: get_all_pages_in_category() and get_pages_in_category() have inconsistent input parsing

git-svn-id: https://perlwikipedia.googlecode.com/svn@355 6dbc9f61-ac24-0410-afbc-9547d35e9fc7