Make get_pages() actually accept an arrayref
get_pages() was advertised as being able to accept an arrayref
of page titles, but it only accepted an array. This checks whether
the next item is an arrayref, and accepts it if so. Otherwise, we
take the rest of the array.

This doubles the number of tests in t/18_get_pages.t, since we
do them once passing page titles as an array, and again passing
them as an arrayref. Since one was always supposed to pass an
arrayref, passing titles as an array may not always be supported.

This also applies the fix to the 3.1 tag.

Resolves issue 106: get_pages() doesn't actually accept an arrayref of titles

git-svn-id: https://perlwikipedia.googlecode.com/svn@376 6dbc9f61-ac24-0410-afbc-9547d35e9fc7