Fix test_image_exists()
This sub attempted to handle an arrayref of images to test, but it
did so in a totally broken way. This disables handling of multiple
images in one query (which was never documented, so isn't a
breaking change) and makes it work right for a single image at a
time. In the future, it would be nice to allow this to handle an
arrayref of images properly.

git-svn-id: https://perlwikipedia.googlecode.com/svn@415 6dbc9f61-ac24-0410-afbc-9547d35e9fc7