Allow changing protocol, host, and path independently in set_wiki()
For people working on secure.wikimedia.org, it makes lots of sense
to allow $bot->set_wiki({path=>"wikipedia/$lang/w"}); rather than
always requiring the protocol and host to be set. Now, set_wiki()
will retain all of protocol, host, and path which are not set. If
they've never been set, the same defaults apply (enwiki, normal
server. The test for this change is in t/01_login.t.

Resolves Issue 130: Allow changing the protocol, host, and path independently in set_wiki()

git-svn-id: https://perlwikipedia.googlecode.com/svn@433 6dbc9f61-ac24-0410-afbc-9547d35e9fc7