Deprecate set_highlimits(), which doesn't do anything
The apihighlimits userright allows an account to retrieve more
results at once than normal. We have a set_highlimits method that
sets an internal flag - but that flag isn't used anywhere. I
believe it was used in the past to decide how many results to
request from the API, but most methods currently ask for 'max',
which means we don't need to know whether we have apihighlimits,
or whether set_highlimits was used. There is no conceivable reason
not to use highlimits if you have the userright, and we do so by
default (by virtue of using 'max').

This deprecates set_highlimits(), and the method will be removed
in a future release.

See issue 147. This is a follow-up to r476 and r484.

git-svn-id: https://perlwikipedia.googlecode.com/svn@489 6dbc9f61-ac24-0410-afbc-9547d35e9fc7